requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyyaml==6.0.1
jinja2==3.1.2
python-multipart==0.0.6
pytest==7.4.3
